fix: remove previous registration links and add new links
